Progress on Rapp and Galvo API.
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@8715 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ plugins/Projector/src/org/micromanager/projector/ProjectionDevice.java
@@ -14,8 +14,8 @@ import java.awt.geom.AffineTransform;
  */
 public interface ProjectionDevice {
    public void displaySpot(double x, double y);
-   public int getWidth();
-   public int getHeight();
+   public double getWidth();
+   public double getHeight();
    public void turnOn();
    public void turnOff();
    public void setRoi(Roi roi, AffineTransform transform);
